correction model liked/disliked, array de string et non string

+++ controllers/sauce.js
@@ -1,20 +1,21 @@
 const Sauce = require('../models/sauce');
+const fs = require('fs'); // « file system » , donne accès aux fonctions qui nous permettent de modifier le système de fichiers, y compris la suppression
 
 exports.createSauce = (req, res, next) => {
   const sauceSchema = JSON.parse(req.body.sauce);
+  console.log(sauceSchema);
   delete sauceSchema._id;
   delete sauceSchema._userId;
   const sauce = new Sauce({
     ...sauceSchema,
-    userId: req.auth.userId,
     imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
-    heat: req.body.heat,
-    // likes: 0,
-    // dislikes: 0,
-    // usersLiked: [""],
-    // usersDisliked: [""],
+    likes: 0,
+    dislikes: 0,
+    usersLiked: [" "],
+    usersDisliked: [" "],
   });
-  sauce.save().then(
+  sauce.save()
+  .then(
     () => {
       res.status(201).json({
         message: 'Object enregistré!'
@@ -22,7 +23,7 @@ exports.createSauce = (req, res, next) => {
     }
   ).catch(
     (error) => {
-      res.status(400).json({
+      res.status(402).json({
         error
       });
     }
@@ -59,20 +60,22 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({_id: req.params.id})
-  .then(
-    () => {
-      res.status(200).json({
-        message: 'Deleted!'
+  Sauce.findOne({ _id: req.params.id})
+      .then(sauce => {
+          if (sauce.userId != req.auth.userId) {
+              res.status(401).json({message: 'Not authorized'});
+          } else {
+              const filename = sauce.imageUrl.split('/images/')[1];
+              fs.unlink(`images/${filename}`, () => {
+                  Sauce.deleteOne({_id: req.params.id})
+                      .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                      .catch(error => res.status(401).json({ error }));
+              });
+          }
+      })
+      .catch( error => {
+          res.status(500).json({ error });
       });
-    }
-  ).catch(
-    (error) => {
-      res.status(400).json({
-        error: error
-      });
-    }
-  );
 };
 
 exports.getAllSauces = (req, res, next) => {
